Downgrade mongo_dart to 0.10.5
The mongo_dart package was downgraded from version 0.10.6 to 0.10.5.
This change also involved moving the dependency from the `dependencies` section to `dev_dependencies` in `pubspec.yaml`, although this was then reverted and it remained a direct main dependency.
The `pubspec.lock` file was updated to reflect this version change and its transitive dependencies.