refactor: Remove toJson override and add private constructor to UserDto
The `toJson()` method in `UserDto` was overriding a method from a superclass or interface. This change removes the `@override` annotation as it's not strictly necessary for Dart classes unless overriding a member from a superclass or interface.

A private constructor `const UserDto._();` was added. This is a common pattern in Dart to prevent instantiation of a class that is meant to be used as a utility class or a class with only static members, though in this case, with a factory constructor present, its direct utility might be related to enforcing the use of the factory constructor for object creation.